These files are for applications.
These files are for applications. The password shall be provided in the resume on the first page. access to them is provided only on a read only basis. Changing and using my code for your own gain or to post as free code is hereby deemed tantamount to copyright infringement and I hall litigate to the last coin I have. I have worked hard for these softwares but I need to show them for application purposes. I might be one guy and broke but I never stop going for my goal, not in my work and not in my battles.